Refactor and add ASCII to end
Refactored the system used to print lines to accomodate incomplete lines
at the end of the file. This is now handled by printLine, which handles
lines whether or not they are "full". The program now reads the bytes
and saves them into a vector and passes this vector to printLine.
Also added a constant for the number of bytes per line so it can be
changed on compile.