Fix grading script
Included the old grading script so that the tests pass. For some reason,
git says that all line feeds will be replaced with carriage return -
line feeds on this commit, but I thought I chose an option on install
that said not do that.
Since there are no formatting issues to be resolved, this will be my
final commit. Locally, all tests pass, 30/30.